Validation and rest of the menu added

+++ App.Library/Interfaces/Repository/IProductRepository.cs
@@ -7,6 +7,6 @@ namespace App.Library.Interfaces.Repository
 {
     public interface IProductRepository : IRepository<ProductModel>
     {
-        void Create(string name, double pricePerKg);
+        void Create(string name, decimal pricePerKg);
     }
 }
